requirements.txt
streamlit>=1.50.0,<1.51.0
plotly>=5.24.0
pandas>=2.2.0
numpy>=1.24.0
scikit-learn>=1.5.0
python-docx>=1.1.2
pdfplumber>=0.11.0
pypdf>=5.0.0
networkx>=3.2.0
rapidfuzz>=3.9.0
streamlit-js-eval>=1.0.0
streamlit-plotly-events>=0.0.6
streamlit-tree-select>=0.0.5
pydantic>=2.8.0
pytest>=8.3.0
